+++ hw/chisel/src/main/scala/snax/streamer/Streamer.scala
@@ -1,5 +1,7 @@
 package snax.streamer
 
+import snax.readerWriter._
+import snax.csr_manager._
 import snax.utils._
 
 import chisel3._
@@ -8,219 +10,153 @@ import chisel3.util._
 // data to accelerator interface generator
 // a vector of decoupled interface with configurable number and configurable width for each port
 class DataToAcceleratorX(
-    params: StreamerParams
+    param: StreamerParam
 ) extends Bundle {
-  val data = MixedVec((0 until params.dataReaderNum).map { i =>
-    Decoupled(UInt(params.fifoWidthReader(i).W))
-  } ++ (0 until params.dataReaderWriterNum).map { i =>
-    Decoupled(UInt(params.fifoWidthReaderWriter(i).W))
+  val data = MixedVec((0 until param.readerNum).map { i =>
+    Decoupled(UInt(param.fifoWidthReader(i).W))
+    // even for reader, from streamer to accelerator
+  } ++ (0 until param.readerWriterNum / 2).map { i =>
+    Decoupled(UInt(param.fifoWidthReaderWriter(i * 2).W))
   })
 }
 
 // data from accelerator interface generator
 // a vector of decoupled interface with configurable number and configurable width for each port
 class DataFromAcceleratorX(
-    params: StreamerParams
+    param: StreamerParam
 ) extends Bundle {
-  val data = MixedVec((0 until params.dataWriterNum).map { i =>
-    Flipped(Decoupled(UInt(params.fifoWidthWriter(i).W)))
-  } ++ (0 until params.dataReaderWriterNum).map { i =>
-    Flipped(Decoupled(UInt(params.fifoWidthReaderWriter(i).W)))
+  val data = MixedVec((0 until param.writerNum).map { i =>
+    Flipped(Decoupled(UInt(param.fifoWidthWriter(i).W)))
+    // oven for writer, from accelerator to streamer
+  } ++ (0 until param.readerWriterNum / 2).map { i =>
+    Flipped(Decoupled(UInt(param.fifoWidthReaderWriter(i * 2 + 1).W)))
   })
-}
-
-// csr related io
-class StreamerCsrIO(
-    params: StreamerParams
-) extends Bundle {
-
-  // configurations interface for a new data operation
-  val loopBounds_i =
-    if (params.ifShareTempAddrGenLoopBounds == true)
-      Vec(1, Vec(params.temporalDimInt, UInt(params.temporalBoundWidth.W)))
-    else
-      MixedVec((0 until params.temporalDimSeq.length).map { i =>
-        Vec(params.temporalDimSeq(i), UInt(params.temporalBoundWidth.W))
-      })
-
-  val temporalStrides_csr_i =
-    if (params.ifShareTempAddrGenLoopBounds == true)
-      Vec(
-        params.dataMoverNum,
-        Vec(params.temporalDimInt, UInt(params.addrWidth.W))
-      )
-    else
-      MixedVec((0 until params.temporalDimSeq.length).map { i =>
-        Vec(params.temporalDimSeq(i), UInt(params.addrWidth.W))
-      })
-
-  val spatialStrides_csr_i =
-    MixedVec((0 until params.spatialDim.length).map { i =>
-      Vec(params.spatialDim(i), UInt(params.addrWidth.W))
-    })
-
-  val ptr_i =
-    Vec(params.dataMoverNum, UInt(params.addrWidth.W))
-
-  // only has transpose function for the data readers
-  val ifTranspose =
-    if (params.hasTranspose) Some(Vec(params.dataReaderNum, Bool())) else None
 }
 
 // data related io
 class StreamerDataIO(
-    params: StreamerParams
+    param: StreamerParam
 ) extends Bundle {
   // specify the interface to the accelerator
   val streamer2accelerator =
-    new DataToAcceleratorX(params)
+    new DataToAcceleratorX(param)
   val accelerator2streamer =
-    new DataFromAcceleratorX(params)
+    new DataFromAcceleratorX(param)
 
   // specify the interface to the TCDM
   // request interface with q_valid and q_ready
   val tcdm_req =
     (Vec(
-      params.tcdmPortsNum,
-      Decoupled(new TcdmReq(params.addrWidth, params.tcdmDataWidth))
+      param.tcdmPortsNum,
+      Decoupled(new TcdmReq(param.addrWidth, param.tcdmDataWidth))
     ))
   // response interface with p_valid
   val tcdm_rsp = (Vec(
-    params.tcdmPortsNum,
-    Flipped(Valid(new TcdmRsp(params.tcdmDataWidth)))
+    param.tcdmPortsNum,
+    Flipped(Valid(new TcdmRsp(param.tcdmDataWidth)))
   ))
 }
 
-// input and output declaration for streamer generator
+/** This class represents the input and output ports of the streamer top module
+  *
+  * @param param
+  *   the parameters class instantiation for the streamer top module
+  */
 class StreamerIO(
-    params: StreamerParams
+    param: StreamerParam
 ) extends Bundle {
 
-  val csr = Flipped(
-    Decoupled(
-      new StreamerCsrIO(
-        params
-      )
-    )
-  )
+  // ports for csr configuration
+  val csr = new SnaxCsrIO(param.csrAddrWidth)
 
+  // ports for data in and out
   val data = new StreamerDataIO(
-    params
+    param
   )
-
-  val busy_o = Output(Bool())
 }
 
 // streamer generator module
 class Streamer(
-    params: StreamerParams,
-    tagName: String = ""
+    param: StreamerParam
 ) extends Module
     with RequireAsyncReset {
-  override val desiredName = tagName + "Streamer"
+
+  override val desiredName = param.tagName + "Streamer"
 
   val io = IO(
     new StreamerIO(
-      params
+      param
     )
   )
 
-  // bandwidth between the tcdm, the fifo and the accelerator should match constrains in Streamer check
+  // --------------------------------------------------------------------------------
+  // ---------------------- csr manager instantiation--------------------------------
+  // --------------------------------------------------------------------------------
 
-  // the tcdmWidth and the fifoWidth in dataMover should match
-  def tcdmWidth = params.dataReaderParams.map(i =>
-    params.tcdmDataWidth * i.tcdmPortsNum
-  ) ++ params.dataWriterParams.map(i =>
-    params.tcdmDataWidth * i.tcdmPortsNum
-  ) ++ params.dataReaderWriterParams.map(i =>
-    params.tcdmDataWidth * i.tcdmPortsNum
-  )
-  def dataMoverWidth = params.dataReaderParams.map(
-    _.fifoWidth
-  ) ++ params.dataWriterParams.map(_.fifoWidth) ++ params.dataReaderWriterParams
-    .map(
-      _.fifoWidth
+  val reader_csr = param.readerParams.map(_.csrNum).reduce(_ + _)
+  val writer_csr = param.writerParams.map(_.csrNum).reduce(_ + _)
+  val reader_writer_csr = param.readerWriterParams.map(_.csrNum).reduce(_ + _)
+
+  // extra one is the start csr
+  val csrNumReadWrite =
+    reader_csr + writer_csr + reader_writer_csr + 1
+
+  // csrManager instantiation
+  val csrManager = Module(
+    new CsrManager(
+      csrNumReadWrite,
+      // 2 ready only csr for every streamer
+      2,
+      param.csrAddrWidth,
+      param.tagName
     )
-  require(tcdmWidth == dataMoverWidth)
-
-  // the fifoWidth in dataMover and the FIFO width should match
-  def fifoWidth =
-    params.fifoWidthReader ++ params.fifoWidthWriter ++ params.fifoWidthReaderWriter
-  require(fifoWidth == dataMoverWidth)
-
-  // accelerator spatialBounds and elementWidth should match
-  def accDataWidth = params.dataReaderParams.map(i =>
-    i.spatialBounds.product * i.elementWidth
-  ) ++ params.dataWriterParams.map(i =>
-    i.spatialBounds.product * i.elementWidth
-  ) ++ params.dataReaderWriterParams.map(i =>
-    i.spatialBounds.product * i.elementWidth
   )
-  require(fifoWidth == accDataWidth)
 
-  def tcdm_read_ports_num =
-    params.dataReaderTcdmPorts.reduceLeftOption(_ + _).getOrElse(0)
-  def tcdm_write_ports_num =
-    params.dataWriterTcdmPorts.reduceLeftOption(_ + _).getOrElse(0)
+  // --------------------------------------------------------------------------------
+  // ---------------------- data reader/writer instantiation--------------------------
+  // --------------------------------------------------------------------------------
 
   // data readers instantiation
   // a vector of data reader generator instantiation with different parameters for each module
-  val data_reader = Seq((0 until params.dataReaderNum).map { i =>
+  val reader = Seq((0 until param.readerNum).map { i =>
     Module(
-      new DataReader(
-        params.dataReaderParams(i),
-        tagName
+      new Reader(
+        param.readerParams(i),
+        param.tagName
       )
     )
   }: _*)
 
   // data writers instantiation
   // a vector of data writer generator instantiation with different parameters for each module
-  val data_writer = Seq((0 until params.dataWriterNum).map { i =>
+  val writer = Seq((0 until param.writerNum).map { i =>
     Module(
-      new DataWriter(
-        params.dataWriterParams(i),
-        tagName
+      new Writer(
+        param.writerParams(i),
+        param.tagName
       )
     )
   }: _*)
 
-  val data_reader_writer = Seq((0 until params.dataReaderWriterNum).map { i =>
+  // data reader_writers instantiation
+  val reader_writer = Seq((0 until param.readerWriterNum / 2).map { i =>
     Module(
-      new DataReaderWriter(
-        params.dataReaderWriterParams(i),
-        tagName
+      new ReaderWriter(
+        param.readerWriterParams(i / 2),
+        param.readerWriterParams(i / 2 + 1),
+        param.tagName
       )
     )
   }: _*)
 
-  // address generation units instantiation
-  // a vector of address generation unit generator instantiation with different parameters for each module
-  val address_gen_unit =
-    if (params.ifShareTempAddrGenLoopBounds == true)
-      Seq((0 until params.dataMoverNum).map { i =>
-        Module(
-          new TemporalAddrGenUnit(
-            params.temporalAddrGenUnitParams(0),
-            tagName
-          )
-        )
-      }: _*)
-    else
-      Seq((0 until params.dataMoverNum).map { i =>
-        Module(
-          new TemporalAddrGenUnit(
-            params.temporalAddrGenUnitParams(i),
-            tagName
-          )
-        )
-      }: _*)
+  // --------------------------------------------------------------------------------
+  // ---------------------- streamer state machine-----------------------------------
+  // --------------------------------------------------------------------------------
 
-  // signals for state transition
-  val config_valid = WireInit(0.B)
-  val streamming_finish = WireInit(0.B)
-
-  val datamover_states = RegInit(VecInit(Seq.fill(params.dataMoverNum)(0.B)))
+  val streamer_config_fire = Wire(Bool())
+  val streamer_finish = Wire(Bool())
+  val streamer_busy = Wire(Bool())
+  val streamer_ready = Wire(Bool())
 
   // State declaration
   val sIDLE :: sBUSY :: Nil = Enum(2)
@@ -233,7 +169,7 @@ class Streamer(
   chisel3.dontTouch(cstate)
   switch(cstate) {
     is(sIDLE) {
-      when(config_valid) {
+      when(streamer_config_fire) {
         nstate := sBUSY
       }.otherwise {
         nstate := sIDLE
@@ -241,7 +177,7 @@ class Streamer(
 
     }
     is(sBUSY) {
-      when(streamming_finish) {
+      when(streamer_finish) {
         nstate := sIDLE
       }.otherwise {
         nstate := sBUSY
@@ -249,308 +185,161 @@ class Streamer(
     }
   }
 
-  io.busy_o := cstate === sBUSY
-
-  config_valid := io.csr.fire && io.csr.fire && io.csr.fire && io.csr.fire
-
   var reader_writer_idx: Int = 0
-  var reader_writer_idx_rw: Int = 0
 
-  for (i <- 0 until params.dataMoverNum) {
-    when(config_valid && cstate === sIDLE) {
-      datamover_states(i) := 1.B
-    }.otherwise {
-      if (i < params.dataReaderNum) {
-        when(data_reader(i).io.data_movement_done) {
-          datamover_states(i) := 0.B
-        }
-      } else {
-        if (i < params.dataReaderNum + params.dataWriterNum) {
-          when(data_writer(i - params.dataReaderNum).io.data_movement_done) {
-            datamover_states(i) := 0.B
-          }
-        } else {
-          reader_writer_idx =
-            (i - params.dataReaderNum - params.dataWriterNum) / 2
-          reader_writer_idx_rw =
-            (i - params.dataReaderNum - params.dataWriterNum) % 2
-          when(
-            data_reader_writer(reader_writer_idx).io
-              .data_movement_done(reader_writer_idx_rw)
-          ) {
-            datamover_states(i) := 0.B
-          }
-        }
-      }
-    }
-  }
+  streamer_config_fire := csrManager.io.csr_config_out.fire
+  streamer_busy := cstate === sBUSY
+  streamer_ready := cstate === sIDLE
 
-  // data streaming finish when all the data mover finished the data movement
-  streamming_finish := !datamover_states.reduce(_ | _)
+  // if every data reader/writer is not busy
+  streamer_finish := !(reader
+    .map(_.io.busy)
+    .reduce(_ || _) || writer
+    .map(_.io.busy)
+    .reduce(_ || _) || reader_writer
+    .map(_.io.readerInterface.busy)
+    .reduce(_ || _) || reader_writer
+    .map(_.io.writerInterface.busy)
+    .reduce(_ || _))
+  dontTouch(streamer_finish)
 
-  io.csr.ready := cstate === sIDLE
+  // --------------------------------------------------------------------------------
+  // -----------------------data movers start-----------------------------------------
+  // --------------------------------------------------------------------------------
 
-  // signals connections for the instantiated modules
-  // TODO: try to use case map to connect...
-  // address generation units csr configuration interface <> streamer IO
-  if (params.ifShareTempAddrGenLoopBounds == true) {
-    for (i <- 0 until params.dataMoverNum) {
-      if (params.stationarity(i) == 1) {
-        for (j <- 0 until params.temporalDimInt) {
-          if (j == 0) {
-            when(io.csr.bits.loopBounds_i(0)(j) === 0.U) {
-              address_gen_unit(i).io.loopBounds_i.bits(j) := 0.U
-            }.otherwise {
-              address_gen_unit(i).io.loopBounds_i.bits(j) := 1.U
-            }
-          } else {
-            address_gen_unit(i).io.loopBounds_i
-              .bits(j) := io.csr.bits.loopBounds_i(0)(j)
-          }
-        }
-      } else {
-        address_gen_unit(
-          i
-        ).io.loopBounds_i.bits := io.csr.bits.loopBounds_i(0)
-      }
-      address_gen_unit(
-        i
-      ).io.loopBounds_i.valid := io.csr.valid
-      address_gen_unit(
-        i
-      ).io.strides_i.bits := io.csr.bits.temporalStrides_csr_i(
-        i
-      )
-      address_gen_unit(
-        i
-      ).io.strides_i.valid := io.csr.valid
-      address_gen_unit(i).io.ptr_i.bits := io.csr.bits.ptr_i(i)
-      address_gen_unit(i).io.ptr_i.valid := io.csr.valid
-    }
-  } else {
-    for (i <- 0 until params.temporalDimSeq.length) {
-      if (params.stationarity(i) == 1) {
-        for (j <- 0 until params.temporalDimSeq(i)) {
-          if (j == 0) {
-            when(io.csr.bits.loopBounds_i(i)(j) === 0.U) {
-              address_gen_unit(i).io.loopBounds_i.bits(j) := 0.U
-            }.otherwise {
-              address_gen_unit(i).io.loopBounds_i.bits(j) := 1.U
-            }
-          } else {
-            address_gen_unit(i).io.loopBounds_i
-              .bits(j) := io.csr.bits.loopBounds_i(i)(j)
-          }
-        }
-      } else {
-        address_gen_unit(
-          i
-        ).io.loopBounds_i.bits := io.csr.bits.loopBounds_i(i)
-      }
-      address_gen_unit(
-        i
-      ).io.loopBounds_i.valid := io.csr.valid
-      address_gen_unit(
-        i
-      ).io.strides_i.bits := io.csr.bits.temporalStrides_csr_i(
-        i
-      )
-      address_gen_unit(
-        i
-      ).io.strides_i.valid := io.csr.valid
-      address_gen_unit(i).io.ptr_i.bits := io.csr.bits.ptr_i(i)
-      address_gen_unit(i).io.ptr_i.valid := io.csr.valid
-    }
-
-  }
-
-  // data reader and data writer <> streamer IO
-  for (i <- 0 until params.dataMoverNum) {
-    if (i < params.dataReaderNum) {
-      data_reader(i).io.spatialStrides_csr_i.bits := io.csr.bits
-        .spatialStrides_csr_i(i)
-      data_reader(
-        i
-      ).io.spatialStrides_csr_i.valid := io.csr.valid
+  // one clock cycle delay for the start signal to store the configuration first
+  for (i <- 0 until param.dataMoverNum) {
+    if (i < param.readerNum) {
+      reader(i).io.start := streamer_config_fire
     } else {
-      if (i < params.dataReaderNum + params.dataWriterNum) {
-        data_writer(
-          i - params.dataReaderNum
-        ).io.spatialStrides_csr_i.bits := io.csr.bits.spatialStrides_csr_i(i)
-        data_writer(
-          i - params.dataReaderNum
-        ).io.spatialStrides_csr_i.valid := io.csr.valid
+      if (i < param.readerNum + param.writerNum) {
+        writer(i - param.readerNum).io.start := streamer_config_fire
       } else {
-        reader_writer_idx =
-          (i - params.dataReaderNum - params.dataWriterNum) / 2
-        reader_writer_idx_rw =
-          (i - params.dataReaderNum - params.dataWriterNum) % 2
-        data_reader_writer(
+        reader_writer_idx = (i - param.readerNum - param.writerNum) / 2
+        reader_writer(
           reader_writer_idx
-        ).io.spatialStrides_csr_i(reader_writer_idx_rw).bits := io.csr.bits
-          .spatialStrides_csr_i(i)
-        data_reader_writer(
+        ).io.readerInterface.start := streamer_config_fire
+        reader_writer(
           reader_writer_idx
-        ).io.spatialStrides_csr_i(reader_writer_idx_rw).valid := io.csr.valid
+        ).io.writerInterface.start := streamer_config_fire
       }
     }
   }
 
-  // the cfg for transpose for only reader
-  val transpose_cfg = RegInit(0.U.asTypeOf(Vec(params.dataReaderNum, Bool())))
+  // --------------------------------------------------------------------------------
+  // ---------------------- csr manager connection----------------------------------------------
+  // --------------------------------------------------------------------------------
 
-  if (params.hasTranspose) {
-    // store the transpose configuration when the cfg is valid
-    when(config_valid === true.B) {
-      transpose_cfg := io.csr.bits.ifTranspose.get
-    }
-    for (i <- 0 until params.dataReaderNum) {
-      data_reader(i).io.ifTranspose.get := transpose_cfg(i)
-    }
+  // connect the csrManager input and streamertop csr req input
+  csrManager.io.csr_config_in.req <> io.csr.req
+
+  // io.csr and csrManager input connection
+  csrManager.io.csr_config_in.rsp <> io.csr.rsp
+
+  // connect the reader/writer ready and csrManager output ready
+  csrManager.io.csr_config_out.ready := streamer_ready
+
+  // add performance counter for streamer
+  val streamerBusy2Idle = WireInit(false.B)
+
+  streamerBusy2Idle := !streamer_busy && RegNext(streamer_busy)
+
+  val performance_counter = RegInit(0.U(32.W))
+  when(streamer_busy) {
+    performance_counter := performance_counter + 1.U
+  }.elsewhen(streamerBusy2Idle) {
+    performance_counter := 0.U
   }
 
-  // data reader and data writer <> address generation units interface
-  for (i <- 0 until params.dataMoverNum) {
-    if (i < params.dataReaderNum) {
-      address_gen_unit(i).io.ptr_o <> data_reader(i).io.ptr_agu_i
-      data_reader(i).io.addr_gen_done := address_gen_unit(i).io.done
-      data_reader(i).io.zeroLoopBoundCase := address_gen_unit(
-        i
-      ).io.zeroLoopBoundCase
-    } else {
-      if (i < params.dataReaderNum + params.dataWriterNum) {
-        address_gen_unit(i).io.ptr_o <> data_writer(
-          i - params.dataReaderNum
-        ).io.ptr_agu_i
-        data_writer(
-          i - params.dataReaderNum
-        ).io.addr_gen_done := address_gen_unit(i).io.done
-        data_writer(
-          i - params.dataReaderNum
-        ).io.zeroLoopBoundCase := address_gen_unit(i).io.zeroLoopBoundCase
-      } else {
-        reader_writer_idx =
-          (i - params.dataReaderNum - params.dataWriterNum) / 2
-        reader_writer_idx_rw =
-          (i - params.dataReaderNum - params.dataWriterNum) % 2
-        address_gen_unit(i).io.ptr_o <> data_reader_writer(
-          reader_writer_idx
-        ).io.ptr_agu_i(reader_writer_idx_rw)
-        data_reader_writer(
-          reader_writer_idx
-        ).io.addr_gen_done(reader_writer_idx_rw) := address_gen_unit(i).io.done
-        data_reader_writer(
-          reader_writer_idx
-        ).io.zeroLoopBoundCase(reader_writer_idx_rw) := address_gen_unit(
-          i
-        ).io.zeroLoopBoundCase
-      }
-    }
+  // connect the performance counter to the first ready only csr
+  csrManager.io.read_only_csr(0) := streamer_busy
+  csrManager.io.read_only_csr(1) := performance_counter
+
+  // store the configuration csr for each data mover when config fire
+  val csrCfgReg = RegInit(VecInit(Seq.fill(csrNumReadWrite)(0.U(32.W))))
+  val csrCfg = Wire(Vec(csrNumReadWrite, UInt(32.W)))
+  when(streamer_config_fire) {
+    csrCfgReg := csrManager.io.csr_config_out.bits
   }
-
-  // data reader and data writer <> accelerator interface
-  // with a queue between each data mover and accelerator data decoupled interface
-
-  val ReaderFifo = Seq(
-    (0 until params.dataReaderNum + params.dataReaderWriterNum).map { i =>
-      if (i < params.dataReaderNum) {
-        Module(
-          new FIFO(
-            params.fifoReaderParams(i).depth,
-            params.fifoReaderParams(i).width,
-            tagName
-          )
-        )
-      } else {
-        Module(
-          new FIFO(
-            params.fifoReaderWriterParams(i - params.dataReaderNum).depth,
-            params.fifoReaderWriterParams(i - params.dataReaderNum).width,
-            tagName
-          )
-        )
-      }
-    }: _*
+  csrCfg := Mux(
+    streamer_config_fire,
+    csrManager.io.csr_config_out.bits,
+    csrCfgReg
   )
 
-  val WriterFifo = Seq(
-    (0 until params.dataWriterNum + params.dataReaderWriterNum).map { i =>
-      if (i < params.dataWriterNum) {
-        Module(
-          new FIFO(
-            params.fifoWriterParams(i).depth,
-            params.fifoWriterParams(i).width,
-            tagName
-          )
-        )
-      } else {
-        Module(
-          new FIFO(
-            params.fifoReaderWriterParams(i - params.dataWriterNum).depth,
-            params.fifoReaderWriterParams(i - params.dataWriterNum).width,
-            tagName
-          )
-        )
-      }
-    }: _*
-  )
+  // --------------------------------------------------------------------------------
+  // ------------------------------------ csr mapping -------------------------------
+  // --------------------------------------------------------------------------------
 
-  var reader_writer_fifo_idx: Int = 0
-  var reader_writer_acc_ports_idx: Int = 0
-  for (i <- 0 until params.dataMoverNum) {
-    if (i < params.dataReaderNum) {
-      ReaderFifo(i).io.in <> data_reader(i).io.data_fifo_o
-      ReaderFifo(i).io.out <> io.data.streamer2accelerator.data(i)
+  // reader
+  var reader_csr_base = 0
+  for (i <- 0 until param.readerNum) {
+    reader_csr_base = param.readerParams
+      .take(i)
+      .map(_.csrNum)
+      .reduceLeftOption(_ + _)
+      .getOrElse(0)
+    reader(i).io.connectCfgWithList(
+      csrCfg.slice(
+        reader_csr_base,
+        reader_csr_base + param.readerParams(i).csrNum
+      )
+    )
+  }
+
+  // writer
+  var writer_csr_base = param.readerParams.map(_.csrNum).sum
+  for (i <- 0 until param.writerNum) {
+    writer_csr_base = param.readerParams.map(_.csrNum).sum + param.writerParams
+      .take(i)
+      .map(_.csrNum)
+      .reduceLeftOption(_ + _)
+      .getOrElse(0)
+    writer(i).io.connectCfgWithList(
+      csrCfg.slice(
+        writer_csr_base,
+        writer_csr_base + param.writerParams(i).csrNum
+      )
+    )
+  }
+
+  // reader_writer
+  var reader_writer_csr_base =
+    param.readerParams.map(_.csrNum).sum + param.writerParams.map(_.csrNum).sum
+  for (i <- 0 until param.readerWriterNum) {
+    reader_writer_csr_base = param.readerParams
+      .map(_.csrNum)
+      .sum + param.writerParams.map(_.csrNum).sum + param.readerWriterParams
+      .take(i)
+      .map(_.csrNum)
+      .reduceLeftOption(_ + _)
+      .getOrElse(0)
+    if (i % 2 == 0) {
+      reader_writer(i / 2).io.readerInterface.connectCfgWithList(
+        csrCfg.slice(
+          reader_writer_csr_base,
+          reader_writer_csr_base + param.readerWriterParams(i).csrNum
+        )
+      )
     } else {
-      if (i < params.dataReaderNum + params.dataWriterNum) {
-        data_writer(
-          i - params.dataReaderNum
-        ).io.data_fifo_i <> WriterFifo(i - params.dataReaderNum).io.out
-        WriterFifo(
-          i - params.dataReaderNum
-        ).io.in <> io.data.accelerator2streamer
-          .data(i - params.dataReaderNum)
-      } else {
-        reader_writer_idx =
-          (i - params.dataReaderNum - params.dataWriterNum) / 2
-        reader_writer_idx_rw =
-          (i - params.dataReaderNum - params.dataWriterNum) % 2
-        if (reader_writer_idx_rw == 0) {
-          data_reader_writer(
-            reader_writer_idx
-          ).io.data_fifo_o <> ReaderFifo(
-            (i - params.dataReaderNum - params.dataWriterNum) / 2 + params.dataReaderNum
-          ).io.in
-          ReaderFifo(
-            (i - params.dataReaderNum - params.dataWriterNum) / 2 + params.dataReaderNum
-          ).io.out <> io.data.streamer2accelerator.data(
-            (i - params.dataReaderNum - params.dataWriterNum) / 2 + params.dataReaderNum
-          )
-        } else {
-          data_reader_writer(
-            reader_writer_idx
-          ).io.data_fifo_i <> WriterFifo(
-            (i - params.dataWriterNum - params.dataReaderNum) / 2 + params.dataWriterNum
-          ).io.out
-          WriterFifo(
-            (i - params.dataWriterNum - params.dataReaderNum) / 2 + params.dataWriterNum
-          ).io.in <> io.data.accelerator2streamer
-            .data(
-              (i - params.dataWriterNum - params.dataReaderNum) / 2 + params.dataWriterNum
-            )
-        }
-      }
+      reader_writer(i / 2).io.writerInterface.connectCfgWithList(
+        csrCfg.slice(
+          reader_writer_csr_base,
+          reader_writer_csr_base + param.readerWriterParams(i).csrNum
+        )
+      )
     }
   }
 
-  // below does data reader and data writer <> TCDM interface
+  // --------------------------------------------------------------------------------
+  // ---------------------- data reader/writer <> TCDM connection-------------------
+  // --------------------------------------------------------------------------------
 
-  // a scala function that flattens the sequenced tcdm ports
-  // returns the list of (i,j,k) in which i is data mover index, j is the TCDM port index in ith data mover,
-  // k is the overall TCDM port index from the TCDM point of view
-  // for instance:
-  // flattenSeq(Seq(2,3)) returns List((0,0,0), (0,1,1), (1,0,2), (1,1,3), (1,2,4))
-  // flattenSeq(Seq(2,2)) returns List((0,0,0), (0,1,1), (1,0,2), (1,1,3))
+  def tcdm_read_ports_num =
+    param.readerTcdmPorts.reduceLeftOption(_ + _).getOrElse(0)
+  def tcdm_write_ports_num =
+    param.writerTcdmPorts.reduceLeftOption(_ + _).getOrElse(0)
+
   def flattenSeq(inputSeq: Seq[Int]): Seq[(Int, Int, Int)] = {
     var flattenedIndex = -1
     val flattenedSeq = inputSeq.zipWithIndex.flatMap { case (size, dimIndex) =>
@@ -559,47 +348,211 @@ class Streamer(
         (dimIndex, innerIndex, flattenedIndex)
       }
     }
-
     flattenedSeq
   }
 
   // data reader <> TCDM read ports
-  val read_flatten_seq = flattenSeq(params.dataReaderTcdmPorts)
+  val read_flatten_seq = flattenSeq(param.readerTcdmPorts)
   for ((dimIndex, innerIndex, flattenedIndex) <- read_flatten_seq) {
     // read request to TCDM
-    io.data.tcdm_req(flattenedIndex) <> data_reader(dimIndex).io.tcdm_req(
+    io.data.tcdm_req(flattenedIndex) <> reader(dimIndex).io.tcdmReq(
       innerIndex
     )
 
     // signals from TCDM responses
-    data_reader(dimIndex).io.tcdm_rsp(innerIndex) <> io.data.tcdm_rsp(
+    reader(dimIndex).io.tcdmRsp(innerIndex) <> io.data.tcdm_rsp(
       flattenedIndex
     )
   }
 
   // data writer <> TCDM write ports
   // TCDM request port bias based on the read TCDM ports number
-  val write_flatten_seq = flattenSeq(params.dataWriterTcdmPorts)
+  val write_flatten_seq = flattenSeq(param.writerTcdmPorts)
   for ((dimIndex, innerIndex, flattenedIndex) <- write_flatten_seq) {
     // write request to TCDM
-    io.data.tcdm_req(flattenedIndex + tcdm_read_ports_num) <> data_writer(
+    io.data.tcdm_req(flattenedIndex + tcdm_read_ports_num) <> writer(
       dimIndex
-    ).io.tcdm_req(innerIndex)
+    ).io.tcdmReq(innerIndex)
   }
 
   // data reader writer <> TCDM read and write ports
-  val read_write_flatten_seq = flattenSeq(params.dataReaderWriterTcdmPorts)
+  val read_write_flatten_seq = flattenSeq(param.readerWriterTcdmPorts)
   for ((dimIndex, innerIndex, flattenedIndex) <- read_write_flatten_seq) {
     // read request to TCDM
     io.data.tcdm_req(
       flattenedIndex + tcdm_read_ports_num + tcdm_write_ports_num
-    ) <> data_reader_writer(dimIndex).io.tcdm_req(
+    ) <> reader_writer(dimIndex).io.readerInterface.tcdmReq(
       innerIndex
     )
 
     // signals from TCDM responses
-    data_reader_writer(dimIndex).io.tcdm_rsp(innerIndex) <> io.data.tcdm_rsp(
-      flattenedIndex + tcdm_read_ports_num + tcdm_write_ports_num
+    reader_writer(dimIndex).io.readerInterface.tcdmRsp(innerIndex) <> io.data
+      .tcdm_rsp(
+        flattenedIndex + tcdm_read_ports_num + tcdm_write_ports_num
+      )
+  }
+
+  // --------------------------------------------------------------------------------
+  // ---------------------- data reader/writer <> accelerator data connection-------
+  // --------------------------------------------------------------------------------
+
+  for (i <- 0 until param.dataMoverNum) {
+    // reader
+    if (i < param.readerNum) {
+      reader(i).io.data <> io.data.streamer2accelerator.data(i)
+    } else {
+      // writer
+      if (i < param.readerNum + param.writerNum) {
+        writer(
+          i - param.readerNum
+        ).io.data <> io.data.accelerator2streamer
+          .data(i - param.readerNum)
+      } else {
+        // reader_writer
+        reader_writer_idx = (i - param.readerNum - param.writerNum) / 2
+        reader_writer(
+          reader_writer_idx
+        ).io.readerInterface.data <> io.data.streamer2accelerator.data(
+          reader_writer_idx + param.readerNum
+        )
+        reader_writer(
+          reader_writer_idx
+        ).io.writerInterface.data <> io.data.accelerator2streamer
+          .data(
+            reader_writer_idx + param.writerNum
+          )
+      }
+    }
+  }
+
+  // --------------------------------------------------------------------------------
+  // ------------------ csr address map header file generation-----------------------
+  // --------------------------------------------------------------------------------
+
+  def genCSRMap(csrBase: Int, param: ReaderWriterParam, tag: String = "") = {
+    var csrMap = "// CSR Mapp for " + tag + "\n"
+    var csrOffset = csrBase
+    // base pointer
+    csrMap = csrMap + "#define BASE_PTR_" + tag + "_0 " + csrOffset + "\n"
+    csrOffset = csrOffset + 1
+    csrMap = csrMap + "#define BASE_PTR_" + tag + "_1 " + csrOffset + "\n"
+    csrOffset = csrOffset + 1
+
+    // spatial bounds
+    for (i <- 0 until param.aguParam.spatialBounds.length) {
+      csrMap =
+        csrMap + "#define " + "S_STRIDE_" + tag + "_" + i + " " + csrOffset + "\n"
+      csrOffset = csrOffset + 1
+    }
+
+    // temporal bounds
+    for (i <- 0 until param.aguParam.temporalDimension) {
+      csrMap =
+        csrMap + "#define " + "T_BOUND_" + tag + "_" + i + " " + csrOffset + "\n"
+      csrOffset = csrOffset + 1
+    }
+
+    // temporal stride
+    for (i <- 0 until param.aguParam.temporalDimension) {
+      csrMap =
+        csrMap + "#define " + "T_STRIDE_" + tag + "_" + i + " " + csrOffset + "\n"
+      csrOffset = csrOffset + 1
+    }
+
+    // channel enable
+    if (param.configurableChannel) {
+      csrMap =
+        csrMap + "#define " + "ENABLED_CHANNEL_" + tag + " " + csrOffset + "\n"
+      csrOffset = csrOffset + 1
+    }
+
+    // byte mask enable
+    if (param.configurableByteMask) {
+      csrMap =
+        csrMap + "#define " + "ENABLED_BYTE_" + tag + " " + csrOffset + "\n"
+      csrOffset = csrOffset + 1
+    }
+
+    csrMap
+  }
+
+  var csrBase = 0
+  var csrMap = ""
+
+  // reader csr configuration
+  for (i <- 0 until param.readerNum) {
+    csrBase = 960 + param.readerParams
+      .take(i)
+      .map(_.csrNum)
+      .reduceLeftOption(_ + _)
+      .getOrElse(0)
+    csrMap = csrMap + genCSRMap(csrBase, param.readerParams(i), "READER_" + i)
+  }
+
+  // writer csr configuration
+  for (i <- 0 until param.writerNum) {
+    csrBase = 960 + param.readerParams.map(_.csrNum).sum + param.writerParams
+      .take(i)
+      .map(_.csrNum)
+      .reduceLeftOption(_ + _)
+      .getOrElse(0)
+    csrMap = csrMap + genCSRMap(csrBase, param.writerParams(i), "WRITER_" + i)
+  }
+
+  // reader_writer csr configuration
+  for (i <- 0 until param.readerWriterNum) {
+    csrBase = 960 + param.readerParams.map(_.csrNum).sum + param.writerParams
+      .map(_.csrNum)
+      .sum + param.readerWriterParams
+      .take(i)
+      .map(_.csrNum)
+      .reduceLeftOption(_ + _)
+      .getOrElse(0)
+    csrMap = csrMap + genCSRMap(
+      csrBase,
+      param.readerWriterParams(i),
+      "READER_WRITER_" + i
     )
   }
+
+  // start csr
+  csrBase = 960 + param.readerParams.map(_.csrNum).sum + param.writerParams
+    .map(_.csrNum)
+    .sum + param.readerWriterParams.map(_.csrNum).sum
+  csrMap = csrMap + "#define STREAMER_START_CSR " + csrBase + "\n"
+
+  // streamer busy csr
+  csrBase = csrBase + 1
+  csrMap = csrMap + "#define STREAMER_BUSY_CSR " + csrBase + "\n"
+
+  // streamer performance counter csr
+  csrBase = csrBase + 1
+  csrMap = csrMap + "#define STREAMER_PERFORMANCE_COUNTER_CSR " + csrBase + "\n"
+
+  println(csrMap)
+
+  val macro_dir = "./generated/header.h"
+  val macro_template =
+    s"""// Copyright 2024 KU Leuven.
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+//
+// Xiaoling Yi <xiaoling.yi@esat.kuleuven.be>
+// This file is generated by Streamer module in hw/chisel to map the CSR address of Streamer automatically, do not modify it manually
+// Generated at ${java.time.Instant.now()}
+
+""" ++ csrMap
+
+  java.nio.file.Files.write(
+    java.nio.file.Paths.get(macro_dir),
+    macro_template.getBytes(java.nio.charset.StandardCharsets.UTF_8)
+  )
+
+}
+
+object StreamerEmitter extends App {
+  emitVerilog(
+    new Streamer(StreamerParam()),
+    Array("--target-dir", "../../target/snitch_cluster/generated")
+  )
 }
