+++ hw/chisel/src/main/scala/snax/xdma/DesignParams/DesignParams.scala
@@ -9,98 +9,6 @@ import chisel3.util.log2Ceil
  *  Design Params is placed all together with companion object to avoid multiple definition of one config & config conflict
  */
 
-// TCDM Params
-
-class TCDMParam(
-    val addrWidth: Int,
-    val dataWidth: Int,
-    val numChannel: Int,
-    val tcdmSize: Int
-)
-
-object TCDMParam {
-  def apply(dataWidth: Int, numChannel: Int, tcdmSize: Int): TCDMParam =
-    new TCDMParam(log2Ceil(tcdmSize) + 10, dataWidth, numChannel, tcdmSize)
-  // By default, the TCDM is 128kB, 64bit data width, 8 channels
-  def apply(): TCDMParam = apply(dataWidth = 64, numChannel = 8, tcdmSize = 128)
-}
-
-// Streamer Params
-
-class AddressGenUnitParam(
-    val spatialBounds: List[Int],
-    val temporalDimension: Int,
-    val addressWidth: Int,
-    val numChannel: Int,
-    val outputBufferDepth: Int,
-    val tcdmSize: Int
-)
-
-object AddressGenUnitParam {
-  def apply(
-      spatialBounds: List[Int],
-      temporalDimension: Int,
-      numChannel: Int,
-      outputBufferDepth: Int,
-      tcdmSize: Int
-  ): AddressGenUnitParam = new AddressGenUnitParam(
-    spatialBounds = spatialBounds,
-    temporalDimension = temporalDimension,
-    addressWidth = log2Ceil(tcdmSize) + 10,
-    numChannel = numChannel,
-    outputBufferDepth = outputBufferDepth,
-    tcdmSize = tcdmSize
-  )
-
-  // The Very Simple instantiation of the Param
-  def apply(): AddressGenUnitParam = apply(
-    spatialBounds = List(8),
-    temporalDimension = 2,
-    numChannel = 8,
-    outputBufferDepth = 8,
-    tcdmSize = 128
-  )
-}
-
-class ReaderWriterParam(
-    spatialBounds: List[Int] = List(8),
-    temporalDimension: Int = 2,
-    tcdmDataWidth: Int = 64,
-    tcdmSize: Int = 128,
-    numChannel: Int = 8,
-    addressBufferDepth: Int = 8,
-    dataBufferDepth: Int = 8,
-    val configurableChannel: Boolean = false,
-    val configurableByteMask: Boolean = false
-) {
-  val aguParam = AddressGenUnitParam(
-    spatialBounds = spatialBounds,
-    temporalDimension = temporalDimension,
-    numChannel = numChannel,
-    outputBufferDepth = addressBufferDepth,
-    tcdmSize = tcdmSize
-  )
-
-  val tcdmParam = TCDMParam(
-    dataWidth = tcdmDataWidth,
-    numChannel = numChannel,
-    tcdmSize = tcdmSize
-  )
-
-  // Data buffer's depth
-  val bufferDepth = dataBufferDepth
-
-  val csrNum =
-    2 + spatialBounds.length + 2 * temporalDimension + (if (configurableChannel)
-                                                          1
-                                                        else
-                                                          0) + (if (
-                                                                  configurableByteMask
-                                                                ) 1
-                                                                else
-                                                                  0)
-}
-
 // AXI Params
 class AXIParam(
     val dataWidth: Int = 512,
@@ -129,97 +37,3 @@ class DMACtrlParam(
     val readerextparam: Seq[DMAExtensionParam] = Seq[DMAExtensionParam](),
     val writerextparam: Seq[DMAExtensionParam] = Seq[DMAExtensionParam]()
 )
-
-class StreamerParam(
-    // data mover params
-    val readerParams: Seq[ReaderWriterParam],
-    val writerParams: Seq[ReaderWriterParam],
-    val readerWriterParams: Seq[ReaderWriterParam],
-
-    // csr manager params
-    val csrAddrWidth: Int,
-    val tagName: String = "Test_"
-) {
-
-  // reader, writer, reader-writer number inferred paramters
-  val readerNum: Int = readerParams.length
-  val writerNum: Int = writerParams.length
-  val readerWriterNum: Int = readerWriterParams.length
-  val dataMoverNum: Int =
-    readerNum + writerNum + readerWriterNum
-
-  // reader, writer, reader-writer tcdm ports inferred parameters
-  val readerTcdmPorts: Seq[Int] = readerParams.map(_.aguParam.numChannel)
-  val writerTcdmPorts: Seq[Int] = writerParams.map(_.aguParam.numChannel)
-  // The tcdm ports for reader-writer, only the even index (reader's) is used
-  // reader and writer share the same tcdm ports
-  val readerWriterTcdmPorts: Seq[Int] =
-    readerWriterParams
-      .map(_.aguParam.numChannel)
-      .zipWithIndex
-      .filter { case (_, index) => index % 2 == 0 }
-      .map(_._1)
-  val tcdmPortsNum: Int =
-    readerTcdmPorts.sum + writerTcdmPorts.sum + readerWriterTcdmPorts.sum
-
-  // inffered parameters for tcdm
-  val addrWidth = readerParams(0).tcdmParam.addrWidth
-  val tcdmDataWidth = readerParams(0).tcdmParam.dataWidth
-
-  // inffered parameters for data fifos
-  val fifoWidthReader: Seq[Int] = readerParams.map(param =>
-    param.aguParam.numChannel * param.tcdmParam.dataWidth
-  )
-  val fifoWidthWriter: Seq[Int] = writerParams.map(param =>
-    param.aguParam.numChannel * param.tcdmParam.dataWidth
-  )
-  val fifoWidthReaderWriter: Seq[Int] = readerWriterParams.map(param =>
-    param.aguParam.numChannel * param.tcdmParam.dataWidth
-  )
-
-  // design time spatial unrolling factors must match the channel number
-  val totalReaderWriterParams =
-    readerParams ++ writerParams ++ readerWriterParams
-  totalReaderWriterParams.foreach { param =>
-    require(
-      param.aguParam.spatialBounds.reduce(_ * _) == param.aguParam.numChannel
-    )
-  }
-
-}
-
-object StreamerParam {
-  def apply() = new StreamerParam(
-    readerParams = Seq(
-      new ReaderWriterParam(temporalDimension = 6),
-      new ReaderWriterParam(temporalDimension = 3)
-    ),
-    writerParams = Seq(new ReaderWriterParam(temporalDimension = 3)),
-    readerWriterParams = Seq(
-      new ReaderWriterParam(
-        temporalDimension = 3,
-        numChannel = 32,
-        spatialBounds = List(32)
-      ),
-      new ReaderWriterParam(
-        temporalDimension = 3,
-        numChannel = 32,
-        spatialBounds = List(32)
-      )
-    ),
-    csrAddrWidth = 32
-  )
-  def apply(
-      readerParams: Seq[ReaderWriterParam],
-      writerParams: Seq[ReaderWriterParam],
-      readerWriterParams: Seq[ReaderWriterParam],
-      csrAddrWidth: Int,
-      tagName: String
-  ) = new StreamerParam(
-    readerParams = readerParams,
-    writerParams = writerParams,
-    readerWriterParams = readerWriterParams,
-    csrAddrWidth = csrAddrWidth,
-    tagName = tagName
-  )
-}
