--- hw/chisel/src/main/scala/snax/xdma/xdmaStreamer/Reader.scala
@@ -1,140 +1,0 @@
-package snax.xdma.xdmaStreamer
-
-import chisel3._
-import chisel3.util._
-
-import snax.utils._
-import snax.xdma.CommonCells._
-import snax.xdma.DesignParams._
-
-// The reader takes the address from the AGU, offer to requestor, and responser collect the data from TCDM and pushed to FIFO packer to recombine into 512 bit data
-
-class Reader(param: ReaderWriterParam, clusterName: String = "unnamed_cluster")
-    extends Module
-    with RequireAsyncReset {
-
-  override val desiredName = s"${clusterName}_xdma_Reader"
-
-  require(
-    param.configurableByteMask == false,
-    "Byte Mask is not supported in Reader"
-  )
-
-  val io = IO(new ReaderIO(param))
-
-  // New Address Generator
-  val addressgen = Module(
-    new AddressGenUnit(
-      param.aguParam,
-      module_name_prefix = s"${clusterName}_xdma_Reader"
-    )
-  )
-
-  // Requestors to send address to TCDM
-  val requestors = Module(
-    new DataRequestors(
-      tcdmDataWidth = param.tcdmParam.dataWidth,
-      tcdmAddressWidth = param.tcdmParam.addrWidth,
-      numChannel = param.tcdmParam.numChannel,
-      isReader = true,
-      module_name_prefix = s"${clusterName}_xdma_Reader"
-    )
-  )
-
-  // Responsors to receive the data from TCDM
-  val responsers = Module(
-    new DataResponsers(
-      tcdmDataWidth = param.tcdmParam.dataWidth,
-      numChannel = param.tcdmParam.numChannel,
-      module_name_prefix = s"${clusterName}_xdma_Reader"
-    )
-  )
-
-  // Output FIFOs to combine the data from the output of responsers
-  val dataBuffer = Module(
-    new ComplexQueueConcat(
-      inputWidth = param.tcdmParam.dataWidth,
-      outputWidth = param.tcdmParam.dataWidth * param.tcdmParam.numChannel,
-      depth = param.bufferDepth
-    ) {
-      override val desiredName = s"${clusterName}_xdma_Reader_DataBuffer"
-    }
-  )
-
-  addressgen.io.cfg := io.aguCfg
-  addressgen.io.start := io.start
-
-  // addrgen <> requestors
-  requestors.io.zip(addressgen.io.addr).foreach {
-    case (requestor, addrgen) => {
-      requestor.in.addr <> addrgen
-    }
-  }
-
-  // enabledChannel & enabledByteMask
-  if (param.configurableChannel) {
-    requestors.io.zip(io.readerwriterCfg.enabledChannel.asBools).foreach {
-      case (requestor, enable) => {
-        requestor.enable := enable
-      }
-    }
-    responsers.io.zip(io.readerwriterCfg.enabledChannel.asBools).foreach {
-      case (responser, enable) => {
-        responser.enable := enable
-      }
-    }
-  } else {
-    requestors.io.foreach(_.enable := true.B)
-    responsers.io.foreach(_.enable := true.B)
-  }
-
-  if (param.configurableByteMask)
-    requestors.io.foreach(_.in.strb := io.readerwriterCfg.enabledByte)
-  else
-    requestors.io.foreach { case requestor =>
-      requestor.in.strb := Fill(requestor.in.strb.getWidth, 1.U)
-    }
-
-  // ReqRsp Link to exchange necessary data
-  requestors.io.zip(responsers.io).foreach {
-    case (requestor, responser) => {
-      requestor.reqrspLink.rspReady.get := responser.reqrspLink.rspReady
-      responser.reqrspLink.reqSubmit := requestor.reqrspLink.reqSubmit.get
-    }
-  }
-
-  // Req & Rsp <> TCDM
-  requestors.io.zip(io.tcdmReq).foreach {
-    case (requestor, tcdmReq) => {
-      requestor.out.tcdmReq <> tcdmReq
-    }
-  }
-  responsers.io.zip(io.tcdmRsp).foreach {
-    case (responser, tcdmRsp) => {
-      responser.in.tcdmRsp <> tcdmRsp
-    }
-  }
-  // Responser <> DataBuffer, Data Link + NearlyFull
-  dataBuffer.io.in.zip(responsers.io).foreach {
-    case (buffer, responser) => {
-      buffer <> responser.out.data
-    }
-  }
-  responsers.io.zip((dataBuffer.io.nearlyFull)).foreach {
-    case (responser, nearlyFull) => {
-      responser.out.dataFifoNearlyFull := nearlyFull
-    }
-  }
-
-  // DataBuffer <> Output
-  dataBuffer.io.out.head <> io.data
-  // Busy Signal
-  io.busy := addressgen.io.busy | (~addressgen.io.bufferEmpty)
-
-  // The debug signal from the dataBuffer to see if AGU and requestor / responser work correctly: It should be high when valid signal at the combined output is low
-  io.bufferEmpty := dataBuffer.io.allEmpty
-}
-
-object ReaderEmitter extends App {
-  println(getVerilogString(new Reader(new ReaderWriterParam)))
-}
