--- hw/chisel/src/main/scala/snax/xdma/xdmaStreamer/Writer.scala
@@ -1,109 +1,0 @@
-package snax.xdma.xdmaStreamer
-
-import chisel3._
-import chisel3.util._
-
-import snax.utils._
-
-import snax.xdma.CommonCells._
-import snax.xdma.DesignParams._
-
-class Writer(param: ReaderWriterParam, clusterName: String = "unnamed_cluster")
-    extends Module
-    with RequireAsyncReset {
-
-  override val desiredName = s"${clusterName}_xdma_Writer"
-
-  val io = IO(new WriterIO(param))
-
-  // New Address Generator
-  val addressgen = Module(
-    new AddressGenUnit(
-      param.aguParam,
-      module_name_prefix = s"${clusterName}_xdma_Writer"
-    )
-  )
-
-  // Write Requestors
-  // Requestors to send address and data to TCDM
-  val requestors = Module(
-    new DataRequestors(
-      tcdmDataWidth = param.tcdmParam.dataWidth,
-      tcdmAddressWidth = param.tcdmParam.addrWidth,
-      numChannel = param.tcdmParam.numChannel,
-      isReader = false,
-      module_name_prefix = s"${clusterName}_xdma_Writer"
-    )
-  )
-
-  val dataBuffer = Module(
-    new ComplexQueueConcat(
-      inputWidth = param.tcdmParam.dataWidth * param.tcdmParam.numChannel,
-      outputWidth = param.tcdmParam.dataWidth,
-      depth = param.bufferDepth
-    ) {
-      override val desiredName = s"${clusterName}_xdma_Writer_DataBuffer"
-    }
-  )
-
-  addressgen.io.cfg := io.aguCfg
-  addressgen.io.start := io.start
-
-  // addrgen <> requestors
-  requestors.io.zip(addressgen.io.addr).foreach {
-    case (requestor, addrgen) => {
-      requestor.in.addr <> addrgen
-    }
-  }
-
-  // enabledChannel & enabledByteMask
-  if (param.configurableChannel)
-    requestors.io.zip(io.readerwriterCfg.enabledChannel.asBools).foreach {
-      case (requestor, enable) => {
-        requestor.enable := enable
-      }
-    }
-  else requestors.io.foreach(_.enable := true.B)
-
-  if (param.configurableByteMask)
-    requestors.io.foreach(_.in.strb := io.readerwriterCfg.enabledByte)
-  else
-    requestors.io.zipWithIndex.foreach {
-      case (requestor, i) => {
-        requestor.in.strb := Fill(requestor.in.strb.getWidth, 1.U)
-      }
-    }
-
-  // Requestor <> TCDM
-  requestors.io.zip(io.tcdmReq).foreach {
-    case (requestor, tcdmReq) => {
-      requestor.out.tcdmReq <> tcdmReq
-    }
-  }
-
-  // Requestor <> DataBuffer, Data Link
-  requestors.io.zip(dataBuffer.io.out).foreach {
-    case (requestor, dataBuffer) => {
-      requestor.in.data.get <> dataBuffer
-    }
-  }
-
-  // DataBuffer <> Input
-  dataBuffer.io.in.head <> io.data
-  // Busy Signal
-  io.busy := addressgen.io.busy | (~addressgen.io.bufferEmpty)
-
-  // The debug signal from the dataBuffer to see if AGU and requestor work correctly: It should be high when valid signal at the combined output is low
-  io.bufferEmpty := addressgen.io.bufferEmpty & dataBuffer.io.allEmpty
-}
-
-object WriterPrinter extends App {
-  println(getVerilogString(new Writer(new ReaderWriterParam)))
-}
-
-object WriterEmitter extends App {
-  emitVerilog(
-    new Writer(new ReaderWriterParam),
-    Array("--target-dir", "generated")
-  )
-}
